+++ docs/apis/apiTag.js
@@ -39,6 +39,20 @@ export default [
                 type: 'Boolean',
                 values: '—',
                 default: '<code>false</code>'
+            },
+            {
+                name: '<code>tabstop</code>',
+                description: 'If should stop when using tab key',
+                type: 'Boolean',
+                values: '—',
+                default: '<code>true</code>'
+            },
+            {
+                name: '<code>disabled</code>',
+                description: 'Disable delete button',
+                type: 'Boolean',
+                values: '—',
+                default: '<code>false</code>'
             }
         ],
         events: [