+++ src/components/tag/Tag.vue
@@ -6,8 +6,9 @@
             <slot></slot>
         </span>
         <a class="tag is-delete"
-            tabindex="0"
             role="button"
+            :tabindex="tabstop ? 0 : false"
+            :disabled="disabled"
             :class="[size, { 'is-rounded': rounded }]"
             @click="close()"
             @keyup.delete.prevent="close()">
@@ -19,12 +20,14 @@
 
         <slot></slot>
 
-        <button v-if="closable"
-            type="button"
+        <a v-if="closable"
+            role="button"
             class="delete is-small"
+            :disabled="disabled"
+            :tabindex="tabstop ? 0 : false"
             @click="close()"
             @keyup.delete.prevent="close()">
-        </button>
+        </a>
     </span>
 </template>
 
@@ -36,7 +39,12 @@
             closable: Boolean,
             type: String,
             size: String,
-            rounded: Boolean
+            rounded: Boolean,
+            disabled: Boolean,
+            tabstop: {
+                type: Boolean,
+                default: true
+            }
         },
         methods: {
             /**
@@ -44,6 +52,8 @@
              * or delete key is pressed.
              */
             close() {
+                if (this.disabled) return
+
                 this.$emit('close')
             }
         }