+++ src/utils/FormElementMixin.js
@@ -12,7 +12,8 @@ export default {
     },
     data() {
         return {
-            isValid: true
+            isValid: true,
+            isFocused: false
         }
     },
     computed: {
@@ -20,14 +21,6 @@ export default {
          * Find parent Field, max 3 levels deep.
          */
         parentField() {
-            // return this.$parent.$data._isField
-            //     ? this.$parent
-            //     : this.$parent.$data._isAutocomplete && this.$parent.$parent.$data._isField
-            //         ? this.$parent.$parent
-            //         : this.$parent.$data._isDatepicker && this.$parent.$parent.$parent.$data._isField
-            //             ? this.$parent.$parent.$parent
-            //             : null
-
             let parent = this.$parent
             for (let i = 0; i < 3; i++) {
                 if (parent && !parent.$data._isField) {
@@ -60,13 +53,20 @@ export default {
          * Focus method that work dynamically depending on the component.
          */
         focus() {
-            if (this.$refs[this.$data._elementRef] === undefined) return
+            if (this.$data._elementRef === undefined) return
 
-            if (!this.$data._isSelect && !this.$data._isAutocomplete && !this.$data._isDatepicker) {
-                this.$nextTick(() => this.$refs[this.$data._elementRef].select())
-            } else {
-                this.$nextTick(() => this.$refs[this.$data._elementRef].focus())
-            }
+            this.$nextTick(() => this.$el.querySelector(this.$data._elementRef).focus())
+        },
+
+        onBlur($event) {
+            this.isFocused = false
+            this.$emit('blur', $event)
+            this.checkHtml5Validity()
+        },
+
+        onFocus($event) {
+            this.isFocused = true
+            this.$emit('focus', $event)
         },
 
         /**
@@ -77,9 +77,7 @@ export default {
         checkHtml5Validity() {
             if (this.$refs[this.$data._elementRef] === undefined) return
 
-            const el = this.$data._isAutocomplete || this.$data._isDatepicker || this.$data._isTimepicker
-                ? this.$refs.input.$refs.input
-                : this.$refs[this.$data._elementRef]
+            const el = this.$el.querySelector(this.$data._elementRef)
 
             let type = null
             let message = null